Create Scope On Fail
Code that adds an 'On Fail' scope to your flow. Copy this and paste this in Power Automate via the 'My clipboard (preview)' option when you add a new action. Simply press Ctrl+v and it should appear.

This code assumes that your main actions are all defined under a scope named 'Main'. Change the 'Filter' action if your scope is named differently.